+++ src/invoke-lambda/_plugin.js
@@ -2,7 +2,7 @@ let read = require('@architect/inventory/src/read')
 let defaultFunctionConfig = require('@architect/inventory/src/defaults/function-config')
 let invocator = require('./')
 
-module.exports = function invokePluginFunction (inventory, { src, payload }, callback) {
+module.exports = function invokePluginFunction ({ inventory, update }, { src, payload }, callback) {
   let params = {
     lambda: {
       src,
@@ -10,7 +10,8 @@ module.exports = function invokePluginFunction (inventory, { src, payload }, cal
       _proxy: true // short circuits sandbox's lambda invocation handler check
     },
     event: payload,
-    inventory
+    inventory,
+    update,
   }
   invocator(params, callback)
 }